增加springboot3.1.1-mybatisplus-redis-java17

+++ springcloud3.1.1-mybatisplus-redis-java17/$module$/src/main/java/$package$/$model$/ctrl/$className$Controller.java
@@ -145,7 +145,7 @@ public class $className$Controller {
         }
         /***}}***/
 
-        int total = $classNameLower$Service.count(queryWrapper);
+        long total = $classNameLower$Service.count(queryWrapper);
         if (total > 0) {
             queryWrapper.lambda().orderByDesc($className$::getId);
